+++ src/lib.rs
@@ -92,5 +92,5 @@ pub mod error;
 pub mod v4;
 pub mod v6;
 
-pub use hickory_proto::error::ProtoError as NameError;
+pub use hickory_proto::ProtoError as NameError;
 pub use hickory_proto::rr::Name;